fix(tools)!: move files around
