+++ .changeset/tools-remove-esbuild.md
@@ -1,4 +1,12 @@
 ---
 "@patternfly/pfe-tools": major
 ---
-❌ Removed esbuild helpers
+Removed and deprecated esbuild and cem/a features
+
+### Breaking Changes
+
+- ❌ Removed esbuild helpers
+- 💱 Moved custom-elements-manifest config helper to `custom-elements-manifest/config.js`
+- 💱 Moved dev-server config helper to `dev-server/config.js`
+- 💱 Moved test-runner config helper to `test-runner/config.js`
+- ⚠️  Deprecated custom-elements-manifest helpers
