+++ elements/custom-elements-manifest.config.js
@@ -1,4 +1,4 @@
-import { pfeCustomElementsManifestConfig } from '@patternfly/pfe-tools/custom-elements-manifest.js';
+import { pfeCustomElementsManifestConfig } from '@patternfly/pfe-tools/custom-elements-manifest/config.js';
 
 export default pfeCustomElementsManifestConfig({
   globs: [
