+++ web-dev-server.config.js
@@ -1,4 +1,4 @@
-import { pfeDevServerConfig } from '@patternfly/pfe-tools/dev-server.js';
+import { pfeDevServerConfig } from '@patternfly/pfe-tools/dev-server/config.js';
 
 export default pfeDevServerConfig({
   tsconfig: 'tsconfig.settings.json',
